ctf-challenge-solutions/requirements.txt
requests
pycryptodome
scapy
